FROMLIST: sched: Initialize CFS task load and usage before placing task on rq
Task load or usage is not currently considered in select_task_rq_fair(),
but if we want that in the future we should make sure it is not zero for
new tasks.

The load-tracking sums are currently initialized using sched_slice(),
that won't work before the task has been assigned a rq. Initialization
is therefore changed to another semi-arbitrary value, sched_latency,
instead.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

(am from https://patchwork.kernel.org/patch/6736681)
Signed-off-by: Juri Lelli <juri.lelli@arm.com>

Change-Id: Ic94c1e767494857e2d2e38eedaf2466063d36ca1