rqbalance: Implement CPU Frequency Limits via CPUFreq API
This commit implements the CPU Frequency Limits functionality
by using the CPUFreq API.
The limits policy is controllable through the rqbalance's
sysfs, with entries named "cluster_freq_vote_max" to vote for
the maximum CPU frequency and "cluster_freq_vote_min" to vote
for the minimum CPU frequency.

The string format to set the minimum or maximum limits is
"NCLUSTER FREQUENCY(KHz)" so, for example, to set a limit of
133MHz on the LITTLE cluster, the string would be "0 133000",
written to either cluster_freq_vote_max or cluster_freq_vote_min.

The vote can be removed by sending "0" (zero) as frequency for
both min or max (in the case of max, zero will be internally
interpreted as UINT_MAX).

Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>