cpuquiet: Add facility to switch from hotplugging to isolation
Some ARM64 SoCs have firmware-managed low power states that
will reduce the power consumption of an online CPU to an order
of microamperes and resuming from these is faster than from a
full hotplug, but these states will kick in only after a short
predefined (in firmware) delay.

To allow using the automated low power states for the AP cores,
artificially isolate CPUs in the scheduler, so that we aren't
executing anything on them and we're sure to give time to the
firmware to switch to these deep states (and back).

This introduces a default option for the cpuquiet governors
for them to be able to initialize with isolation instead of
hotplugging and also gives to the user the possibility to
force a switch to either hotplug or isolation in run-time by
writing on the sysfs node "use_sched_isolation".

Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>